2.后台Python/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==6.1.0
certifi==2025.7.14
click==8.2.1
configparser==7.2.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Pillow==10.4.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.8.0
python-multipart==0.0.12
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
zhipuai==2.1.5.20250716
